From Uwe Ligges<ligges@statistik.tu-dortmund.de> 6/6/2014 ligges@statistik.tu-dortmund.de
After internal discussion:

Assuming that SUNDIALS is really BSD-3-clause (please confirm):


1. BSD-3-clause is known to be compatible with GPL-3.

2. You LICENSE template is perhaps not appropriate, as it
only relates to the foreign code.

3. What we need is one license for the package which could be GPL-3 given your LICENSE.note and our interpretation is correct, but we do not give legal advice here.

So we think the only reasonable solution is to have the package under
  GPL-3 only.  I.e.,

     License: GPL-3

  and then remove files LICENSE and LICENSE.note.
But be careful to kep all copyright statement and really list all the authors of the material included in your package.

git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@8780 1feff8c3-07ed-0310-ac33-dd36852eb9cd